% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latamverse-package.R, R/latamverse.R
\docType{package}
\name{Latamverse}
\alias{Latamverse}
\alias{Latamverse-package}
\title{Latamverse: Latin American Data via RESTful APIs and Curated Datasets}
\usage{
Latamverse()
}
\value{
An object of class "latamverse_info" containing package names and versions.
}
\description{
{
This package brings together a comprehensive collection of R packages providing access to APIs functions and curated datasets from Argentina, Brazil, Chile, Colombia, and Peru. It includes real-time and historical data through public RESTful APIs (Nager.Date, World Bank API, REST Countries API, and country-specific APIs) and extensive curated collections of open datasets covering economics, demographics, public health, environmental data, political indicators, social metrics, and cultural information.
}

This function creates an object containing information about the API packages
included in the Latamverse metapackage and their respective versions.
}
\details{
Latamverse: Latin American Data via RESTful APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Latin American Data via RESTful APIs and Curated Datasets.
}
\examples{
\dontrun{
# Display Latamverse package information
Latamverse()

# Store the information in an object
info <- Latamverse()
print(info)
}
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/latamverse}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
