% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latamverse.R
\name{print.latamverse_info}
\alias{print.latamverse_info}
\title{Print method for latamverse_info objects}
\usage{
\method{print}{latamverse_info}(x, ...)
}
\arguments{
\item{x}{An object of class "latamverse_info"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for latamverse_info objects
}
