% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latamverse-package.R, R/latamverse.R
\docType{package}
\name{Latamverse}
\alias{Latamverse}
\alias{Latamverse-package}
\title{Latamverse: A Metapackage for Latin American Countries via RESTful APIs and Curated Datasets}
\usage{
Latamverse()
}
\value{
Invisibly returns the names of the loaded packages.
}
\description{
{
This metapackage brings together a comprehensive collection of R packages providing access to APIs functions and curated datasets from Argentina, Brazil, Chile, Colombia, and Peru. It includes real-time and historical data through public RESTful APIs (Nager.Date, World Bank API, REST Countries API, and country-specific APIs) and extensive curated collections of open datasets covering economics, demographics, public health, environmental data, political indicators, social metrics, and cultural information.
}

This function displays a formatted list of the API packages included
in the Latamverse metapackage and their respective versions.
}
\details{
Latamverse: A Metapackage for Latin American Countries via RESTful APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Metapackage for Latin American Countries via RESTful APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/latamverse}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
